Fibonacci example, bug fixes, set instruction
Created a fibonacci example (maybe buggy and/or may not work properly). Fixed a bug that caused blank lines to not be ignored by the pre-processor. Added a "set instruction" --- which can be used to set a variable. The current implementation is buggy and will need to be improved upon down the line. There is now a "parse_string" function.